chore: move layer creation logic to a new file
